refactor: split cli to allow programmatic use
